requirements.txt
black
python-vlc
wxpython
